error correction in time detection tests
